% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rule.R
\name{add_new_rule}
\alias{add_new_rule}
\title{Create a custom rule for internal use}
\usage{
add_new_rule(name, path = ".")
}
\arguments{
\item{name}{Name(s) of the rule. Cannot contain white space.}

\item{path}{Path to package or project root.}
}
\description{
This function creates a YAML file with the placeholder text to define a new
rule. The file is stored in \code{flir/rules/custom}. You need to create the
\code{flir} folder with \code{setup_flir()} if it doesn't exist.

If you want to create a rule that users of your package will be able to
access, use \code{export_new_rule()} instead.
}
