% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{expect_s4_class_linter}
\alias{expect_s4_class_linter}
\title{Require usage of \code{expect_s4_class(x, k)} over \code{expect_true(is(x, k))}}
\usage{
expect_s4_class_linter
}
\value{
The name of the linter
}
\description{
See \url{https://lintr.r-lib.org/reference/expect_s4_class_linter}.
}
