% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rule.R
\name{export_new_rule}
\alias{export_new_rule}
\title{Create a custom rule for external use}
\usage{
export_new_rule(name, path = ".")
}
\arguments{
\item{name}{Name(s) of the rule. Cannot contain white space.}

\item{path}{Path to package or project root.}
}
\description{
This function creates a YAML file with the placeholder text to define a new
rule. The file is stored in \code{inst/flir/rules} and will be available to users
of your package if they use \code{flir}.

To create a new rule that you can use in the current project only, use
\code{add_new_rule()} instead.
}
