% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{for_loop_index_linter}
\alias{for_loop_index_linter}
\title{Block usage of for loops directly overwriting the indexing variable}
\usage{
for_loop_index_linter
}
\value{
The name of the linter
}
\description{
See \url{https://lintr.r-lib.org/reference/for_loop_index_linter}.
}
