% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-linters.R
\name{list_linters}
\alias{list_linters}
\title{Get the list of linters in \code{flir}}
\usage{
list_linters(path = ".")
}
\arguments{
\item{path}{A valid path to a file or a directory. Relative paths are
accepted. Default is \code{"."}.}
}
\value{
A character vector
}
\description{
Get the list of linters in \code{flir}
}
\examples{
list_linters(".")
}
