% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{nested_ifelse_linter}
\alias{nested_ifelse_linter}
\title{Block usage of nested \code{ifelse()} calls}
\usage{
nested_ifelse_linter
}
\value{
The name of the linter
}
\description{
See \url{https://lintr.r-lib.org/reference/nested_ifelse_linter}.
}
