% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{nzchar_linter}
\alias{nzchar_linter}
\title{Require usage of nzchar where appropriate}
\usage{
nzchar_linter
}
\value{
The name of the linter
}
\description{
See \url{https://lintr.r-lib.org/reference/nzchar_linter}.
}
