% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_flir}
\alias{setup_flir}
\title{Setup flir}
\usage{
setup_flir(path = ".")
}
\arguments{
\item{path}{Path to package or project root.}
}
\value{
Imports files necessary for \code{flir} to work but doesn't return any
value in R.
}
\description{
This stores the default rules and internal files in \code{inst/flir}. It also
imports \code{sgconfig.yml} that is used by \code{ast-grep}. This file must live at the
root of the project and cannot be renamed.
}
