% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{update_flir}
\alias{update_flir}
\title{Update the \code{flir} setup}
\usage{
update_flir(path = ".")
}
\arguments{
\item{path}{Path to package or project root.}
}
\value{
Can add new files in the \code{flir/rules} folder, doesn't return anything.
}
\description{
When \code{flir} is updated, it can ship new built-in rules or update existing
ones. \code{update_flir()} will automatically add those new rules to the
\code{flir/rules/builtin} folder. Custom rules stored in \code{flir/rules/custom}
are not affected.
}
\examples{
\dontrun{
  update_flir()
}
}
