% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{vector_logic_linter}
\alias{vector_logic_linter}
\title{Enforce usage of scalar logical operators in conditional statements}
\usage{
vector_logic_linter
}
\value{
The name of the linter
}
\description{
See \url{https://lintr.r-lib.org/reference/vector_logic_linter}.
}
