% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{which_grepl_linter}
\alias{which_grepl_linter}
\title{Require usage of grep over which(grepl(.))}
\usage{
which_grepl_linter
}
\value{
The name of the linter
}
\description{
See \url{https://lintr.r-lib.org/reference/which_grepl_linter}.
}
