% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_rule.R
\name{add_new_rule}
\alias{add_new_rule}
\title{Create a custom rule}
\usage{
add_new_rule(name, path = ".")
}
\arguments{
\item{name}{Name of the rule. Cannot contain white space.}

\item{path}{Path to package or project root.}
}
\value{
Creates a new file but doesn't return anything.
}
\description{
Create a custom rule
}
